Requirement: #6020 - Image Fusion (Phase 1) Ticket: #6732 - Fusion: asynchronous loading '''Schema/Configuration Changes: None'''  * Added interactive progress bar graphics to core viewer.

+++ ImageViewer/Imaging/IdentityVoiLinearLut.cs
@@ -135,7 +135,7 @@ namespace ClearCanvas.ImageViewer.Imaging
 		/// </remarks>
 		public double WindowWidth
 		{
-			get { return _maxInputValue - _minInputValue + 1; }
+			get { return (double) _maxInputValue - _minInputValue + 1; }
 		}
 
 		/// <summary>
